All the recent changes introduced by Tom and Roger replicated in PM version
git-svn-id: https://svn.code.sf.net/p/electricdss/code/trunk@2130 d8739450-1e93-4ef4-a0af-c327d92816ff

+++ Parallel_Version/Source/TCP_IP/TCP_IP.pas
@@ -716,7 +716,7 @@ begin
          npts := ActiveLSObject.NumPoints;
      SetLength(time, 1, npts) ;
      for i:=0 to npts-1 do
-      time[0,i] := double(i)*sinterfal;
+      time[0,i] := i*sinterfal;
      x_label:='Time (s)';
     end;
